Implement gene validator service, remove unused files
