+++ package.json
@@ -76,6 +76,7 @@
     "react-router-dom": "^4.2.2",
     "react-table": "^6.7.4",
     "react-tabs": "^2.1.1",
+    "request": "^2.83.0",
     "rethinkdb": "^2.3.3",
     "sbgnml-to-cytoscape": "^4.0.1",
     "serve-favicon": "^2.4.5",
