--- src/server/gene-validator/tokenize.js
@@ -1,5 +1,0 @@
-const tokenize = function(input) {
-  return input.split(/\r?\n/).filter(String);
-};
-
-module.exports = tokenize;
