--- test/mapping-test.js
@@ -1,20 +1,0 @@
-// mapping HGNC symbol to HGNC ID
-const chai = require('chai');
-const expect = chai.expect;
-const mapping = require('../src/server/gene-query/mapping.js');
-
-describe('mapping HGNC symbol to HGNC ID', function() {
-  it('it should return a HGNC ID', function() {
-    expect(mapping('AFF4')).to.deep.equal('HGNC:17869');
-    expect(mapping('HLTF')).to.deep.equal('HGNC:11099');
-    expect(mapping('HCFC1')).to.deep.equal('HGNC:4839');
-    expect(mapping('NCOA1')).to.deep.equal('HGNC:7668');
-    expect(mapping('MTA2')).to.deep.equal('HGNC:7411');
-    expect(mapping('MBD1')).to.deep.equal('HGNC:6916');
-    expect(mapping('JMJD6')).to.deep.equal('HGNC:19355');
-    expect(mapping('MTA3')).to.deep.equal('HGNC:23784');
-    expect(mapping('GTF3C4')).to.deep.equal('HGNC:4667');
-    expect(mapping('BRD4')).to.deep.equal('HGNC:13575');
-    expect(mapping('ASH1L')).to.deep.equal('HGNC:19088');
-  });
-});
